+++ src/data/+mag/Instrument.m
@@ -80,10 +80,25 @@ classdef Instrument < handle & matlab.mixin.Copyable & matlab.mixin.CustomDispla
                 filters
             end
 
-            hasScience = this.HasScience;
-
+            % Crop science.
             this.cropScience(filters{:});
-            this.cropToMatch(HadScience = hasScience);
+
+            % TODO: it should be smarter and convert to the max/min of all
+            % filters.
+            filter = filters{1};
+
+            % Crop events.
+            if ~isempty(this.Events)
+                this.Events = this.Events.crop(filter);
+            end
+
+            % Adjust metadata.
+            if ~isempty(this.Metadata)
+                this.Metadata.Timestamp = this.TimeRange(1);
+            end
+
+            % Filter HK.
+            this.HK.crop(filter);
         end
 
         function cropScience(this, filters)
@@ -106,7 +121,7 @@ classdef Instrument < handle & matlab.mixin.Copyable & matlab.mixin.CustomDispla
             end
         end
 
-        function cropToMatch(this, startTime, endTime, options)
+        function cropToMatch(this, startTime, endTime)
         % CROPTOMATCH Crop metadata, events and HK based on science
         % timestamps or specified timestamps.
 
@@ -114,7 +129,6 @@ classdef Instrument < handle & matlab.mixin.Copyable & matlab.mixin.CustomDispla
                 this (1, 1) mag.Instrument
                 startTime (1, 1) datetime = this.TimeRange(1)
                 endTime (1, 1) datetime = this.TimeRange(2)
-                options.HadScience (1, 1) logical = true
             end
 
             timePeriod = timerange(startTime, endTime, "closed");
@@ -127,11 +141,6 @@ classdef Instrument < handle & matlab.mixin.Copyable & matlab.mixin.CustomDispla
             % Adjust metadata.
             if ~isempty(this.Metadata)
                 this.Metadata.Timestamp = startTime;
-            end
-
-            % If there already wasn't any science, do not crop the HK.
-            if ~options.HadScience
-                return;
             end
 
             % Filter HK.