+++ resources/release-notes.md
@@ -1,3 +1,9 @@
+## App
+
+- (All) Add "Health" tab to display health check results
+
 ## Software
 
-- (IMAP) Allow running IMAP analysis without any input data pattern
+- (IMAP) Add health checks for secondary currents and voltages, temperatures, saturation, missed ITFs, and activation
+- (All) `mag.Instrument/crop` method crops data based on input timerange, not science timerange
+- (All) `mag.version` should not return a `datetime`-like string